+++ neuralnetwork_foundations/algorithms/__init__.py
@@ -1,1 +1,3 @@
-from linear_regression import LinearRegression
+from .LinearRegression import LinearRegression
+from .kMeans import kMeans
+from .kNearestNeighbor import kNearestNeighbor
